backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Scheduler
apscheduler==3.10.4

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Platform APIs (optional, use if not implementing manually)
tweepy==4.14.0  # For X/Twitter

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
